A new routine called update_halo has been included.
This routine works in two steps.
First the halos are cleared (ie: all the bergs are deleted from cells in the halos).
Second, the icebergs near the the edge of the processor are copied into the halos of the other processor.

The code compiles, but has not yet been tested.

Also, I am not sure that the icebergs which are being moved into the halos of the other processor are being sent to the correct grid list. This will be the next update.