Added a new type called bond, which corresponds to bonds between icebergs
Created a submodule which initially established bonds between icebergs close to one another (for testing purposes)

Created a submodule which checks whether the bonds are health, and can also count the total number of bounds.

There are still some bugs, and everyone is not quite working, but it is now time to go home.