Icebergs bonds can now move from one processor to another. After the bond has moved, it is reconstructed on the other processor.
I still need to write a scheme which locates the partner bergs on the other processor. However, this will be simple, since the berg stores the address of the partner berg in the bond, and moves it to the other processor.

I have also introduced a namelist parameter called max_bonds, which is the maximum amount of bonds that one iceberg can have when it is moved from one processor to another.

This all seems to be working for now.