Added bond interactions.
The bond interactions only take place is the distance between the two icebergs is greater than R1+R2.
At the moment, this is a spring force, which does not depend on the mass of the icebergs. This requires that all bonds are broken before the iceberg becomes very small, or else the berg will become unstable. This should be revisited.

The iceberg bond interactions have not been tested yet. Before testing I need to find the bug in the verlet code.