lat_old, lon_old, uvel_old, vvel_old do not need to be passed between processors, or passed to the IO.
I have removed these, which should speed things up a little.

I have also changed buffer_width so that it is no longer a parameter. This allows us to define buffer_width in terms of the namelist variable max_bonds. This assigns the correct amount of memory for passing bonds. At a later stage we will perhaps adjust the size of the buffer depending on the amount of bonds.