Two bugs in the iceberg interactions fixed:
1) Damping coefitients must be multiplied by dt before entering the matrix A (which solves for the updated acceleration.

2) P_21 was not being correctly undated when applying the damping to perpendicular motion. This has been corrected.

Also, the perpendicular damping is set to 4 times less than the parralel. This is just a guess for now, and needs to be investigated further.

Also, a flag has been added whihc allows the user to use icebergs which only feel interaction forces, and do not feel the environment.

Fixing these two bugs allows the model to run with more than 4 bonded icebergs. More testing is needed to see if this is sufficiently robust