There were a few bugs in the iceberg tipping sheme:
1) The icebergs should tip if their shortest horizonal length (W) is small compared to the iceberg depth  (model was using the large horizonal length L ) - this bug was copied from Bigg et al 1997
2) There was an error in copying from Bigg et al 1997. The tipping scheme used the iceberg draft instead of the iceberg thickness.
3) In the Weeks and Mellor paper, there was an error which meant that the plus sign in front of the 58.32 should have been a minus sign

Instead of correcting these errors a flag has been added so that a different rolling criteria is used where icebergs roll when W/Th>tip_parameter. The tip parameter can be added at runtime, or else is calculated directly from the iceberg and ocean density. This is the formula given in Burton et al 2012, equation 27, or equivolently it is the same as the Weeks and Mellor 1978 result, using an iceberg of constant density.

When run with the standard configuration, the old (incorrect) rolling scheme is used, and the code does not change the answers