A bug has been fixed, which prevented the model from reproducing across different layouts.
The model was copying icebergs from the restart file onto two different processors. This was because when an iceberg is placed directly on the edge of a grid cell, which is positioned on the edge of a processor, then the iceberg was determined to be in both cells, and was thus placed on both processors.

This has been fixed by editing the routine sum_sign_dot_prod4, so that it is asymtric, including points positioned on the south/west boundaries, and not including points on the north/east.
Note: This has only been done for the sum_sign_dot_prod4, and not for the sum_sign_dot_prod5 routine, which is used for icebregs near the northpole.

Two different tests have also been written to make sure that the correct amount of icebergs are read from the restart file.