A piece of code has been added to the routine which handels spreading mass to ocean. The new code reassigns mass which should be spread onto land points (masked points). The unused mass is assigned to the other grid around the iceberg in proportion to the percent of the iceberg which is in that grid cell. This piece of code completes the hexagonal spreading, and allows the hexagonal spreading to interact with boundaries.
For the square spreading, the code should not make any difference, because all of the mass is already distributed (with the center grid cell taking all the mass which would have otherwise been put onto a masked cell. At some point we may want to change the way this works (and make it work more like the hexagonal case). However, for now, we leave it as it was, so that the answers should not change.