Code has been added to allow for non periodic domains, for periodic domains which are not of length 360, and also for grids which are not lat lon grids.
The periodicity of the domain (in the x-direction) is in imput variable. For non-periodic domains, this should be set to a very large number.
A flag has been added which signals whether the grid_is_latlon. The various metric terms have been adjusted according to this flag.

Period domains in the y-direction have not yet been implemented.

The code has not yet been tested, but it does compile.