The length of the domain, Lx, and the flag grid_is_latlon have been included in the gridded_icebergs type, which means that these variables do not have to be passed around as explicitly. This neatens the code.
A flag has been added to allow the code to run on an f-plane. The reference latitude for the f_plane is an input parameter, lat_ref.

The option to change omega has been removed. No rotation can be achieved by setting the lat_ref to zero and using the f-plane (this is the default for when grid_is_latlon is off)

The model has been set to default to an f-plane when grid_is_latlon is off.

The periodicity in the x-direction is set by Lx. When grid_is_latlon is set to False, and Lx=360 (default), then Lx is set to a very large number which makes the model not periodic.