Fixing is_point_in_cell routine for points near cell edge
A bug was found in the routine is_point_in_cell. For icebergs with a longitude very close to the edge of a the cell, the modulo arithemtic inside sum_sign_prod4 rounds off the berg position to be exactly on the edge of the cell. However, the one edge of the cell actually belows to the next cell over. Therefore after roundoff, the berg is found not to be inside the cell.

However, the 'crude bounds' inside is_point_in_cell mean that the adjacent cell (which the berg will be in after roundoff) does not make it through. This means that the icebergs is not found in either cell, and is not found at all.

This problem has been solved by weakening the crude bound test. A tolorance has been added to the crude bounds inside the routine is_point_in_cell. This tol=0.1 avoids excluding points which are very near to the edge of the cell.

This is a bit of an hack solution since the tolorance is in longitude coordinates. However it works for now.