Removed the flag save_static_berg_field_in_restart
The flag save_static_berg_field_in_restart has been removed since it was frustrating to have to remember to include this flag every time in order to get a useable restart file.
Instead, the field static_berg is included in the restart file whenever there exists an iceberg which has static_berg > 0.

This means that in Production runs where this field is not being used, the field will not be saved.