docs: update CPP ifdef logic to set "file version"
doxygen is mishandling the CPP #ifdef logic and mishandling differently
on different platforms (possible dependence on the version of /lib/cpp).
The form of #ifdef A #else B #endif allowed doxygen to see two
declarations of a variable.  Refactoring the logic to #ifdef A #endif
seems to fix the problem.