Avoids type ambiguity for variable/function
Gnu compiler was complaining with "Derived type 'buffer' is used as an actual argument ..."
for a one character variable being passed to class(*) in FMS2. It turns out the compiler
was seeing the function 'buffer' in the "use" statements above. Renaming the variable
avoids the ambiguity.

Thanks to @marshallward for figuring out the cause.